Fix the TF --> TFLM sync
 * added the new files to be sync'd to ci/tflite_files.txt
 * manually ran the sync script: ci/sync_from_upstream_tf.sh
 * update BUILD files to match changes from http://cl/488654028

BUG=http://b/259507701
