+++ ci/tflite_files.txt
@@ -1,11 +1,11 @@
 tensorflow/lite/builtin_op_data.h
 tensorflow/lite/builtin_ops.h
 tensorflow/lite/context_util.h
-tensorflow/lite/c/common.cc
 tensorflow/lite/core/api/error_reporter.cc
 tensorflow/lite/core/api/flatbuffer_conversions.cc
 tensorflow/lite/core/api/op_resolver.cc
 tensorflow/lite/core/api/tensor_utils.cc
+tensorflow/lite/core/c/common.cc
 tensorflow/lite/kernels/internal/quantization_util.cc
 tensorflow/lite/kernels/kernel_util.cc
 tensorflow/lite/schema/schema_utils.cc
@@ -16,6 +16,9 @@ tensorflow/lite/core/api/error_reporter.h
 tensorflow/lite/core/api/flatbuffer_conversions.h
 tensorflow/lite/core/api/op_resolver.h
 tensorflow/lite/core/api/tensor_utils.h
+tensorflow/lite/core/c/builtin_op_data.h
+tensorflow/lite/core/c/c_api_types.h
+tensorflow/lite/core/c/common.h
 tensorflow/lite/kernels/internal/common.h
 tensorflow/lite/kernels/internal/compatibility.h
 tensorflow/lite/kernels/internal/portable_tensor_utils.h
