+++ tensorflow/lite/kernels/internal/common.h
@@ -411,7 +411,7 @@ LUTPopulate(float input_scale, int32_t input_zero_point, float output_scale,
   for (int32_t val = minval; val <= maxval; ++val) {
     const float dequantized = input_scale * (val - input_zero_point);
     const float transformed = transform(dequantized);
-    const float rescaled = std::round(transformed * inverse_scale);
+    const float rescaled = TfLiteRound(transformed * inverse_scale);
     const int32_t quantized =
         static_cast<int32_t>(rescaled + output_zero_point);
     lut_uint8[static_cast<uint8_t>(static_cast<T>(val))] = static_cast<uint8_t>(
